web/api/serializers.py
from rest_framework import serializers
from .models import Coin, Block

class BlockSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Block
        fields = '__all__'


class CoinSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Coin
        fields = '__all__'
