web/requirements.txt
Django
psycopg2
djangorestframework
pytz
drf-yasg==1.20.0